Bug Fix
I have already submitted this project. I was not able to properly test
this app so there were a few bugs beyond the alarm service JSONQuery.
Strings being casted to floats for comparison were not valid due to the
$ symbol. I removed the '$'
When notification gives data back to the activity the footer is removed
because we are not loading any more data.
I added a service content creator to my dummy content creator to fake
data from Zappos so I could get beyond the web call in the service to
test the rest of the app.